Refactor DefaultNavigation
